# openapi.yaml
openapi: 3.0.0
info:
  x-skaffolder-id-project: 62782285c97c70218d654808
  title: j2ee
  description: You can obtain JWT token from login API
  version: 0.1
servers:
  - url: http://localhost:3000/api
    description: Local server for testing
paths:
...
  /user:
    post:
      x-skaffolder-id: 62782286c97c70218d654815
      x-skaffolder-name: create
      x-skaffolder-id-resource: 62782285c97c70218d654812
      x-skaffolder-resource: User
      x-skaffolder-crudAction: create
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION create
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION create
      tags:
        - User
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                mail:
                  type: string
                name:
                  type: string
                password:
                  type: string
                phone:
                  type: number
                username:
                  type: string
    get:
      x-skaffolder-id: 62782286c97c70218d654818
      x-skaffolder-name: list
      x-skaffolder-id-resource: 62782285c97c70218d654812
      x-skaffolder-resource: User
      x-skaffolder-crudAction: list
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION list
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION list
      tags:
        - User
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
  "/user/{id}":
    delete:
      x-skaffolder-id: 62782286c97c70218d65481c
      x-skaffolder-name: delete
      x-skaffolder-id-resource: 62782285c97c70218d654812
      x-skaffolder-resource: User
      x-skaffolder-crudAction: delete
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION delete
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION delete
      tags:
        - User
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      x-skaffolder-id: 62782286c97c70218d654817
      x-skaffolder-name: get
      x-skaffolder-id-resource: 62782285c97c70218d654812
      x-skaffolder-resource: User
      x-skaffolder-crudAction: get
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION get
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION get
      tags:
        - User
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    post:
      x-skaffolder-id: 62782286c97c70218d654819
      x-skaffolder-name: update
      x-skaffolder-id-resource: 62782285c97c70218d654812
      x-skaffolder-resource: User
      x-skaffolder-crudAction: update
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION update
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION update
      tags:
        - User
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
  "/user/findByusername/{key}":
    get:
      x-skaffolder-id: 62782384c97c70218d65481f
      x-skaffolder-name: findByusername
      x-skaffolder-id-resource: 62782285c97c70218d654812
      x-skaffolder-resource: User
      x-skaffolder-crudAction: findByusername
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION findByusername
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /findByusername/{key}
      x-skaffolder-roles: []
      summary: CRUD ACTION findByusername
      tags:
        - User
      parameters:
        - name: username
          x-skaffolder-type: String
          in: path
          description: Id of the resource username to search
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      x-skaffolder-id: 62782285c97c70218d654812
      x-skaffolder-id-db: 62782285c97c70218d654809
      x-skaffolder-id-entity: 62782285c97c70218d65480a
      x-skaffolder-url: /user
      x-skaffolder-type: User
      properties:
        _id:
          type: string
          x-skaffolder-required: true
        mail:
          type: string
          x-skaffolder-id-attr: 62782285c97c70218d65480d
          x-skaffolder-type: String
        name:
          type: string
          x-skaffolder-id-attr: 62782285c97c70218d65480e
          x-skaffolder-type: String
        password:
          type: string
          x-skaffolder-id-attr: 62782285c97c70218d65480c
          x-skaffolder-type: String
          x-skaffolder-required: true
        phone:
          type: number
          x-skaffolder-id-attr: 62782686c97c70218d654826
          x-skaffolder-type: Number
        username:
          type: string
          x-skaffolder-id-attr: 62782285c97c70218d65480b
          x-skaffolder-type: String
          x-skaffolder-required: true
      x-skaffolder-relations: 
      required:
        - _id
        - password
        - username
...